Give nice error message instead of traceback if trying to set system/user vm symlink over existing non-symlink file/directory. Fixes bug #140926
svn path=/projects/java-config-2/trunk/; revision=2711

+++ src/java_config/EnvironmentManager.py
@@ -214,8 +214,11 @@ class EnvironmentManager:
         if not os.path.isdir(sym_dir):
             os.makedirs(sym_dir)
 
-        if os.path.islink(target):
-            os.remove(target)
+        if os.path.lexists(target):
+            if os.path.islink(target):
+                os.remove(target)
+            else:
+                raise InvalidConfigError(target)
 
         os.symlink('/usr/lib/jvm/'+vm.name(),target)
